Edited list of places and form

+++ main.py
@@ -1,4 +1,4 @@
-from flask import request, flash, abort, render_template, url_for
+from flask import request, flash, abort, jsonify, render_template, url_for
 from flask_login import login_required, current_user, login_user, logout_user, LoginManager  # , UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
 # from flask_mail import Mail
@@ -7,8 +7,10 @@ import json
 # import os
 # from token import generate_confirmation_token, confirm_token
 # from check_email import send_email
-from forms import RegistrationForm
-from database.models import app, db, User, Place
+from forms import UserSchema
+from database.models import *
+from marshmallow import ValidationError
+from sqlalchemy.exc import IntegrityError
 
 login_manager = LoginManager()
 login_manager.init_app(app)
@@ -32,6 +34,37 @@ MAIL_PASSWORD = os.environ['LeisureGuru12345']
 '''
 
 
+def error_handler(func):
+    def wrapper(*args, **kwargs):
+        # print("error_handler")
+        try:
+            # result = 0
+            if 0 == len(kwargs):
+                result = func()
+            else:
+                result = func(**kwargs)
+            if result[1] >= 400:
+                return {
+                    "code": result[1],
+                    "message": result[0]
+                }, result[1]
+            else:
+                return result
+        except ValidationError as err:
+            # print(err.messages)
+            return {"code": 412,
+                    "message": err.messages
+                    }, 412
+        except IntegrityError as err:
+            # print(err.args)
+            return {"code": 409,
+                    "message": err.args
+                    }, 409
+
+    wrapper.__name__ = func.__name__
+    return wrapper
+
+
 @login_manager.user_loader
 def load_user(user_id):
     return User.query.get(user_id)
@@ -46,7 +79,7 @@ def home():
     return "Home page :)"
 
 
-@app.route("/signup", methods=['POST', 'GET'])
+@app.route("/register", methods=['POST', 'GET'])
 def signup():
     if request.method == 'POST':
         if request.is_json:
@@ -103,38 +136,17 @@ def confirm_email(token):
 '''
 
 
-@app.route('/register', methods=['GET', 'POST'])
+@app.route('/signup', methods=['GET', 'POST'])
+@error_handler
 def register():
-    if request.method == 'POST':
-        if request.is_json:
-            register_form = RegistrationForm(request.form)
-            if register_form.validate_on_submit():
-                new_user = User(
-                    first_name=register_form.first_name.data,
-                    last_name=register_form.last_name.data,
-                    email=register_form.email.data,
-                    birth_date=register_form.birth_date.data,
-                    password1=generate_password_hash(register_form.password.data)
-                )
-                find_email = User.query.filter_by(email=new_user.email).first()
-                if find_email is not None:
-                    flash("Email is already used", "error")
-                else:
-                    db.session.add(new_user)
-                    db.session.commit()
-                    return {"id": new_user.id,
-                            "email": new_user.email}
-                '''
-                token = generate_confirmation_token(new_user.email)
-                confirm_url = url_for('new_user.confirm_email', token=token, _external=True)
-                html = render_template('activate.html', confirm_url=confirm_url)
-                subject = "Please confirm your email"
-                send_email(new_user.email, subject, html)
-                '''
-                login_user(new_user)
-
-                flash('A confirmation email has been sent via email.', 'success')
-            return "Home"  # redirect(url_for("main.home"))
+    if request.method == 'POST' and request.is_json:
+        data_user = UserSchema().load(request.json)
+        new_user = User(**data_user)
+        new_user.status = True
+        db.session.add(new_user)
+        db.session.commit()
+        login_user(new_user)
+        return jsonify(UserSchema().dump(new_user)), 201
     return "Register"  # render_template('/register', form=form)
 
 
@@ -157,19 +169,21 @@ def login():
     return "Login :)"
 
 
-@app.route("/user/<int:user_id>", methods=['GET', 'DELETE', 'POST'])
+@app.route("/profile/<int:user_id>", methods=['GET', 'DELETE', 'POST', 'PUT'])
 @login_required
 def user(user_id):
-    user_to_work = User.query.get_or_404(user_id)
-    if request.method == 'GET':
+    # user_to_work_data = request.get_json()
+    user_to_work = User.query.filter_by(id=user_id).first()
+    if request.method == 'GET' and user_to_work != []:
+        print("Got", user_id)
         user_to_work.status = False
         # db.session.add(user_to_work)
         db.session.commit()
         logout_user()
         # db.session.pop('id', None)
         # db.session.pop('email', None)
-    elif request.method == 'DELETE':
-        if user_id == current_user.id:
+    elif request.method == 'DELETE' and user_to_work != []:
+        if user_to_work.id == current_user.id:
             db.session.delete(user_to_work)
             db.session.commit()
             flash("Success")
@@ -181,12 +195,8 @@ def user(user_id):
 
 @app.route("/homepage", methods=['GET'])
 # @login_required
-def place():
-    if request.method == 'GET':
-        places = load_all_places()
-        # we need to remake it to work with FE correctly
-        return json.dumps([p.as_dict() for p in Place.query.all()])
-    return "Place"
+def homepage():
+    return json.dumps([p.as_dict() for p in Place.query.all()])
 
 
 if __name__ == "__main__":
